requirements.txt
streamlit==1.32.0

langchain==0.1.16
langchain-core==0.1.53
langchain-community==0.0.36
langchain-text-splitters==0.0.2

langchain-groq==1.1.2
groq>=0.5.0,<1.0.0

faiss-cpu==1.13.2
sentence-transformers==3.0.1

python-dotenv==1.0.1
numpy==1.26.4